chore: Fix types
