+++ editor.planx.uk/src/routes/serviceFeedback.tsx
@@ -2,11 +2,11 @@ import gql from "graphql-tag";
 import { FEEDBACK_SUMMARY_FIELDS } from "lib/feedback";
 import { compose, mount, NotFoundError, route, withData } from "navi";
 import { FeedbackLog } from "pages/FlowEditor/components/FeedbackLog/FeedbackLog";
+import { Feedback } from "pages/FlowEditor/components/FeedbackLog/types";
 import { useStore } from "pages/FlowEditor/lib/store";
 import React from "react";
 
 import { client } from "../lib/graphql";
-import { Feedback } from "./feedback";
 import { makeTitle } from "./utils";
 
 const serviceFeedbackRoutes = compose(
